Unify findProjectConfigByName into config.ts and thread config through keepalive
Move findProjectConfigByName from mag.ts to config.ts as a shared exported
helper, eliminating duplicated project name/repo-tail matching logic.
Thread LudicsFullConfig through getSortedReadyCandidates, maybeQueueProposals,
and maybeFillEmptySlots to avoid redundant loadConfigSync() calls in the
keepalive hot path.
